+++ src/device/device.mstest/Tests/Asserts.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace cc.isr.VI.Device.Tests;
 /// <summary>
 /// Test manager for <see cref="VisaSessionBase"/> and <see cref="SubsystemBase"/> Tests.