+++ src/vi/k2600.ttm/legacy.mstest/Tests/Asserts.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace cc.isr.VI.Tsp.K2600.Ttm.Legacy.Tests;
 
 /// <summary>   Asserts for testing the legacy driver. </summary>