+++ src/vi/k2600.ttm/mstest/Tests/Firmware/Asserts_.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace cc.isr.VI.Tsp.K2600.Ttm.Tests.Firmware;
 internal static partial class Asserts
 {