+++ src/vi/k2600.ttm/mstest/Tests/Firmware/FirmwareTests.cs
@@ -65,7 +65,9 @@ public class FirmwareTests
         // reported in the base class
         Console.WriteLine( $"{this.TestContext?.FullyQualifiedTestClassName}: {DateTime.Now} {System.TimeZoneInfo.Local}" );
         Console.WriteLine( $"\t{typeof( cc.isr.Visa.Gac.Vendor ).Assembly.FullName}" );
-        Console.WriteLine( $"\t{cc.isr.Visa.Gac.GacLoader.LoadedImplementation?.Location}." );
+        if ( cc.isr.Visa.Gac.GacLoader.LoadedImplementation is not null )
+            // the implementation is loaded after the device opens and then stays loaded. 
+            Console.WriteLine( $"\t{cc.isr.Visa.Gac.GacLoader.LoadedImplementation?.Location}." );
         Console.WriteLine( $"\tTesting {typeof( cc.isr.VI.Tsp.K2600.Ttm.Meter ).Assembly.FullName}" );
 
         // create an instance of the Serilog logger.